Updated test_queue.py to use pytest and added more tests.

+++ WindNinja-Server/windninja_server/windninjaqueue/queue.py
@@ -12,12 +12,12 @@ _directories = {
         "queue" : ""
     }
 
-def set_Queue(config, initialize):
+def set_Queue(config, initialize=False):
     global _directories
     _directories["queue"]  = config["datastore"]
 
     global _mode
-    _mode = QueueMode[config["mode"]]
+    _mode = QueueMode[config.get("mode", "disabled")]
 
     if _mode == QueueMode.immediate:
        global wn
@@ -25,14 +25,14 @@ def set_Queue(config, initialize):
        wn.PYTHON_EXECUTABLE = config["windninja_wrapper"]["executable"]
        wn.WN_WRAPPER = config["windninja_wrapper"]["script"]
        wn.WN_WRAPPER_OPTIONS = config["windninja_wrapper"]["options"]
-       
+
     if initialize:
         os.makedirs(_directories["queue"])
-        
+
 def enqueue(id, reset_existing=False):
     if _mode == QueueMode.disabled:
-        return 
-    
+        return
+
     name = "{0}.{1}".format(id, QueueStatus.pending.name)
     file = os.path.join(_directories["queue"], name)
 
@@ -63,7 +63,7 @@ def enqueue(id, reset_existing=False):
             raise Exception(message)
 
 
-def dequeue(id):
+def dequeue(id, **kwargs):
     #just skip out if disabled
     if _mode == QueueMode.disabled:
         return
@@ -71,22 +71,26 @@ def dequeue(id):
     item = _find_item(id)
     if item is None:
         raise KeyError("Item with id {} not found in queue".format(id))
-        
-    update_queue_item_status(id, QueueStatus.complete)
+
+    update_queue_item_status(id, QueueStatus.complete, **kwargs)
+
 
 def update_queue_item_status(id, status, data=None):
     if _mode == QueueMode.disabled:
-        return 
+        return
 
     if type(status) is not QueueStatus:
         raise TypeError("Status is not of type <QueueStatus>")
-    
+
     try:
         existing_file = _find_item(id)
+        if existing_file is None:
+            raise KeyError("Item with id {} not found in queue".format(id))
+
         base = os.path.splitext(existing_file)[0]
         new_file = "{0}.{1}".format(base, status.name)
         os.rename(existing_file, new_file)
-        
+
         if data:
             with open(new_file, "at") as f:
                 f.write(data)
@@ -96,10 +100,7 @@ def update_queue_item_status(id, status, data=None):
         raise KeyError("Item with id {} not found in queue".format(id))
 
     except OSError as osex:
-        raise Exception("Queue update os error\n{}\n".format(str(osex)), data) 
-
-    except Exception as ex:
-        raise Exception("Queue update error: {}".format(str(ex))) 
+        raise Exception("Queue update os error\n{}\n".format(str(osex)), data)
 
 def find_items_by_status(status):
     # get the files by status pattern
@@ -108,7 +109,13 @@ def find_items_by_status(status):
     files = glob.glob(file_pattern)
 
     # return a dictionary with id and created time
-    return [{"id":os.path.splitext(os.path.basename(item))[0], "created":os.path.getctime(item)} for item in files]
+    return [
+        {
+            "id": os.path.splitext(os.path.basename(item))[0],
+            "created": os.path.getctime(item)
+        }
+        for item in files
+    ]
 
 def _find_item(id):
     name_pattern = "{}.*".format(id)
